Разделил schemas.py на файлы в папке src/api/schemas/, заполнил _init_.py файл, проверил что за счет него импорты в остальных файлах не сломались, эндроинты API работают, бот работает
