+++ src/api/schemas/__init__.py
@@ -1,0 +1,17 @@
+from .admin import AdminUserRequest  # noqa
+from .analytics import Analytic  # noqa
+from .base import RequestBase, ResponseBase  # noqa
+from .categories import CategoryRequest, CategoryResponse  # noqa
+from .external_site_user import ExternalSiteUser, ExternalSiteUserRequest  # noqa
+from .health_check import BotStatus, CommitStatus, DBStatus, HealthCheck  # noqa
+from .notification import (  # noqa
+    ErrorsSending,
+    FeedbackFormQueryParams,
+    InfoRate,
+    Message,
+    MessageList,
+    TelegramNotificationRequest,
+    TelegramNotificationUsersGroups,
+    TelegramNotificationUsersRequest,
+)
+from .tasks import TaskRequest, TaskResponse  # noqa
