+++ src/api/schemas/base.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel, Extra
+
+
+class ResponseBase(BaseModel):
+    """Базовый класс для модели ответа."""
+
+    class Config:
+        from_attributes = True
+
+
+class RequestBase(BaseModel):
+    """Базовый класс для модели запроса."""
+
+    class Config:
+        extra = Extra.forbid
