+++ src/api/schemas/external_site_user.py
@@ -1,0 +1,35 @@
+from pydantic import Field, field_validator
+
+from src.api.schemas.base import RequestBase
+from src.core.db.models import ExternalSiteUser
+
+
+class ExternalSiteUserRequest(RequestBase):
+    """Класс модели запроса для ExternalSiteUser."""
+
+    id: int = Field(...)
+    id_hash: str = Field(..., max_length=256)
+    first_name: str | None = Field(None, max_length=64)
+    last_name: str | None = Field(None, max_length=64)
+    email: str = Field(..., max_length=48)
+    specializations: list[int] | None = None
+
+    def to_orm(self) -> ExternalSiteUser:
+        return ExternalSiteUser(
+            id=self.id,
+            id_hash=self.id_hash,
+            email=self.email,
+            first_name=self.first_name,
+            last_name=self.last_name,
+            specializations=self.specializations,
+        )
+
+    @field_validator("specializations", mode="before")
+    def specializations_str_validation(cls, value: str):
+        if not isinstance(value, str):
+            return value
+        try:
+            new_value = [int(value) for value in value.split(", ")]
+            return new_value
+        except ValueError:
+            raise ValueError("Для передачи строки с числами в поле specializations " 'используйте формат: "1, 2, 3" ')
